- Percentiles are now more precisely defined using doubles - Percentiles in JsonSource are now normalized with respect to *total* order volume, which makes directly comparing bid and ask percentiles feasible

+++ src/main/java/c3po/bitstamp/BitstampOrderBookSource.java
@@ -19,11 +19,11 @@ public abstract class BitstampOrderBookSource extends AbstractTickable implement
 	protected final CircularArrayList<ServerSnapshot> buffer;
 	protected boolean isEmpty = false;
 	
-	protected final int[] percentiles;
+	protected final double[] percentiles;
 	
 	private final List<OutputSignal> signals;
 	
-	public BitstampOrderBookSource(long timestep, long interpolationTime, int[] percentiles) {
+	public BitstampOrderBookSource(long timestep, long interpolationTime, double[] percentiles) {
 		super(timestep);
 		
 		this.interpolationTime = interpolationTime;
